fix: css is changed to consisten break point usage and duplicated avatarFallBack function is moved a place
