+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/sectionWrapper.tsx
@@ -16,15 +16,15 @@ type SectionWrapperProps = {
 
 const SectionWrapper = ({ title, description, fullWidth, className, action, children }: SectionWrapperProps) => {
   return (
-    <section className="flex flex-col gap-4 border-b py-6 lg:py-8 first:pt-4 xl:flex-row">
-      <div className="flex w-full flex-col gap-3 xl:max-w-xs">
+    <section className="flex flex-col gap-4 border-b py-6 lg:py-8 first:pt-4 lg:flex-row">
+      <div className="flex w-full flex-col gap-3 lg:max-w-xs">
         <div className="flex w-full flex-col gap-1">
           <h2 className="text-base flex items-center gap-2">{title}</h2>
           <div className="w-full text-sm text-muted-foreground">{description}</div>
         </div>
         {action}
       </div>
-      <div className={cn("grow", !fullWidth && "max-w-none xl:max-w-xl", className)}>{children}</div>
+      <div className={cn("grow", !fullWidth && "max-w-none lg:max-w-xl", className)}>{children}</div>
     </section>
   );
 };
